# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevMind AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/scans:
    get:
      operationId: listSecurityScans
      tags: [security]
      summary: List all security scans
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityScan"
    post:
      operationId: createSecurityScan
      tags: [security]
      summary: Submit code for security analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityScanInput"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityScan"
  /security/scans/{id}:
    get:
      operationId: getSecurityScan
      tags: [security]
      summary: Get a specific security scan result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityScan"
        "404":
          description: Not found
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflow]
      summary: List all generated workflows
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflow]
      summary: Generate a workflow from natural language description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Workflow generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflow]
      summary: Get a specific workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          description: Not found

  # Codebase Knowledge AI
...
components:
...
  schemas:
...
    SecurityScanInput:
      type: object
      required: [code, language]
      properties:
        code:
          type: string
          description: Source code to analyze
        language:
          type: string
          description: Programming language (e.g. python, javascript, java)
        filename:
          type: ["string", "null"]
          description: Optional filename for context
    SecurityScan:
      type: object
      required: [id, code, language, status, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        language:
          type: string
        filename:
          type: ["string", "null"]
        status:
          type: string
          description: "pending | complete | error"
        severityScore:
          type: ["number", "null"]
          description: 0-10 severity score
        vulnerabilities:
          type: ["string", "null"]
          description: JSON array of vulnerability findings
        summary:
          type: ["string", "null"]
          description: AI-generated summary
        createdAt:
          type: string
          format: date-time
...
    WorkflowInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
          description: Natural language description of the automation
    Workflow:
      type: object
      required: [id, description, status, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        status:
          type: string
          description: "pending | complete | error"
        pipelineJson:
          type: ["string", "null"]
          description: JSON representing the pipeline steps
        diagramJson:
          type: ["string", "null"]
          description: JSON representing the visual flow diagram nodes/edges
        createdAt:
          type: string
          format: date-time

    # Codebase
...